# continuation attachments