--- a/mats/patch-compile-0-f-t-f
+++ b/mats/patch-compile-0-f-t-f
@@ -327,8 +327,8 @@
   record.mo:Expected error in mat r6rs-records-procedural: "attempt to apply non-procedure spam".
 ! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure n>".
 ! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure n>".
-! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure pcons>".
-! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure pcons>".
+! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure>".
+! record.mo:Expected error in mat r6rs-records-procedural: "incorrect number of arguments to #<procedure>".
 ! record.mo:Expected error in mat r6rs-records-procedural: "incorrect argument count in call (ccons 1)".
 ! record.mo:Expected error in mat r6rs-records-procedural: "incorrect argument count in call (ccons 1 2 3)".
 ! record.mo:Expected error in mat r6rs-records-procedural: "incorrect argument count in call (n (+ z 7) w "what?")".