--- a/mats/record.ms
+++ b/mats/record.ms
@@ -8853,7 +8853,7 @@
            (make-foo 3))))
     `(let ([ctr 0])
        (letrec ([g0 (lambda (new) (lambda (q) (new q (begin (set! ctr (#2%+ 1 xtr)) ctr))))])
-         (#3%$make-record-constructor-descriptor ',record-type-descriptor?  #f g0 'define-record-type)
+         (#3%$value (#3%$make-record-constructor-descriptor ',record-type-descriptor?  #f g0 'define-record-type))
          (#3%$record ',record-type-descriptor? 3 (begin (set! ctr (#2%+ 1 xtr)) ctr)))))
   (equivalent-expansion?
     (parameterize ([optimize-level 3] [enable-cp0 #t] [#%$suppress-primitive-inlining #f])
@@ -8872,7 +8872,7 @@
            (make-foo 3))))
     `(let ([ctr 0])
        (letrec ([g0 (lambda (new) (lambda (q) (new q (begin (set! ctr (#3%+ 1 xtr)) ctr))))])
-         (#3%$make-record-constructor-descriptor ',record-type-descriptor?  #f g0 'define-record-type)
+         (#3%$value (#3%$make-record-constructor-descriptor ',record-type-descriptor?  #f g0 'define-record-type))
          (#3%$record ',record-type-descriptor? 3 (begin (set! ctr (#3%+ 1 xtr)) ctr)))))
   (error? ; invalid uid
     (let ()