--- a/release_notes/release_notes.stex
+++ b/release_notes/release_notes.stex
@@ -58,6 +58,23 @@ Online versions of both books can be found at
 %-----------------------------------------------------------------------------
 \section{Functionality Changes}\label{section:functionality}
 
+\subsection{Continuation attachments (9.5.3)}
+
+The new \scheme{call-setting-continuation-attachment},
+\scheme{call-getting-continuation-attachment},
+\scheme{call-consuming-continuation-attachment}, and
+\scheme{current-continuation-attachments} procedure support
+efficiently recording and querying information on a continuation, but
+without unnecessarily extending continuation---that is, without
+interfering with tail-call behavior.
+
+As part of this addition, compilationin safe mode avoids moving an
+expression that produces multiple values out of a context that
+requires a single value (so a run-time error is preserved). More
+generally, optimization in all modes preserves observable non-tail
+behavior, and it is more directly observable using continuation
+attachments.
+
 \subsection{Unicode Basic Multilingual Plane console I/O in Windows (9.5.3)}
 
 Console I/O now supports characters from the Unicode Basic