@mflatt 2018-07-26
I've updated the PR to change the way relative-tail calls are handled under `call-setting-continuation-attachment` in a non-tail position. Instead of allocating an attachment-moving wrapper as a closure, the relative-tail call itself is adjusted to perform the move.

The change saves about 10-20% in the simulated contract example or in this variant of "(non-tail) recur using attachments":
```
   (let loop ([j 0])
     (cond
      [(= j N) 'done]
      [else
       (let loop ([i 0])
         (cond
          [(= i M) 0]
          [else (call-setting-continuation-attachment
                 i
                 (lambda ()
                   (sub1 (loop (add1 i)))))]))
       (loop (add1 j))]))
```

It also seems like a generally better approach, although it requires some care to ensure that the new info flag for a `call` or `mvcall` form isn't lost in various compiler passes.