@mflatt 2020-05-20
Rebased and folded in #377 plus subsequent improvements.